+++ CHANGELOG.md
@@ -1,21 +1,29 @@
-## New features
+## Improvements
+
+- The option to crop unreasonably long texts in the tooltip has been moved from the profile sanitization to the tooltip options. The text is no longer cropped when a profile is received but when the text is being displayed. The option is still enabled by default and can be disabled in the settings. If you still see cropped text after disabling the option it means the profile was cropped with the previous system. Delete the profile from the directory to download it again.
+- Emotes starting with a comma `,` are now handled like emotes starting with a `'s`, removing the space inserted by the game after the player's name. — [Ticket #100](https://wow.curseforge.com/projects/total-rp-3/issues/100)
+
+
+## Previous change-log, for version 1.2.8
+
+### New features
 
 - Added option to show custom player icons in names in the chat — [Ticket #91](https://wow.curseforge.com/projects/total-rp-3/issues/91)
 
 ![Custom player icons in chat](https://totalrp3.info/documentation/changelogs/1_2_8_icons.png)
 
-## Added patch 7.2 resources
+### Added patch 7.2 resources
 
 - Added 996 new icons from patch 7.2 (for a total of 17,671 icons available.)
 - Added 19 new images from patch 7.2 to insert in your profiles (for a total of 449 images available).
 - Added 124 new musics from patches 7.1 and 7.2 (for a total of 2,868 musics available. Note that the previous music list contained many duplicates that have been cleaned up).
 
-## Bug fixes
+### Bug fixes
 
 - Fixed an error with the keyboard bindings. — [Ticket #89](https://wow.curseforge.com/projects/total-rp-3/issues/89)
 - Fixed an error in the Prat module with non-customized names. — [Ticket #95](https://wow.curseforge.com/projects/total-rp-3/issues/95)
 
-## Other improvements
+### Other improvements
 
 - Removed the quotes surrounding the "currently" text in the tooltip (IC and OOC). — [Ticket #88](https://wow.curseforge.com/projects/total-rp-3/issues/88)
 - The profile sanitization option will now also crop unreasonably long texts. — [Ticket #92](https://wow.curseforge.com/projects/total-rp-3/issues/92)